requirements.txt
streamlit 
transformers 
pillow
git+https://github.com/huggingface/transformers.git
torch
torchvision
numpy==1.*